Add arms, interventions and outcomes to an existing xml document for upload to clinicaltrials.gov

These functions add arms, interventions, primary and secondary outcomes as well as conditions and keywords to an xml document created using the create_ctxml() function. Calls to these functions should not be assigned to an object.

add_arm(ctxml, arm_label, arm_type, arm_desc)

add_intervention(ctxml, int_name, int_type, int_desc, arm_label)

add_pr_outcome(ctxml, name, time, description)

add_sec_outcome(ctxml, name, time, description)

add_condition(ctxml, condition)

add_keyword(ctxml, keyword)

Arguments

ctxml
A xml document generated from the create_ctxml() function

arm_label
Label assigned to arm of clinical trial. Arm means a pre-specified group or subgroup of participant(s) in a clinical trial assigned to receive specific intervention(s) (or no intervention).

arm_type
Either Experimental, Active comparator, Placebo Comparator, Sham Comparator, No Intervention, or Other.

arm_desc
Description of the arm.

int_name
Name of the intervention. For a drug, it is the generic name.

int_type
Drug, Device, Biological/Vaccine, Procedure/Surgery, Radiation, Behavioural, Genetic, Dietary Supplement, Combination Product, Diagnostic Test, or Other.

int_desc
Other details about the intervention not included in name.

name
Name of outome measure.

time
Time point(s) at which the measurement is assessed.

description
Other details about the outcome measure not included in the name

condition
MeSH term for condition being studied in the trial, or Focus of the Study

keyword
Words or phrases that best describe the protocol. Keywords help users find studies in the database.

Value

A xml document

Details

add_arm(): Adds an xml nodespace containing information about the arm name, type and description to the xml document.

add_intervention(): Adds an xml nodespace containing information about the intervention name, type, description and arm it is associated with to the xml document.

add_pr_outcome(): Adds an xml nodespace containing information about the outcome name, time frame for measurement and additional descriptive details to the xml document.

add_sec_outcome(): Adds an xml nodespace containing information about the outcome name, time frame for measurement and additional descriptive details to the xml document.

add_condition(): Adds an xml nodespace containing a MeSH term for the condition being studied in the trial, or Focus of the Study to the xml document.

add_keyword(): Adds an xml nodespace containing a Words or phrases that best describe the protocol. Keywords help users find studies in the database to the xml document.

Examples

# NOT RUN {
 add_arm(ctxml = ctxml,
         arm_label = "Standard",
         arm_type =  "Active Comparator",
         arm_desc = "Manual upload to registry")
# }

# NOT RUN {
add_intervention(ctxml = ctxml,
                 int_type =  "Device",
                 int_name = "Registry entry",
                 int_desc = "The usual way to enter to the registry",
                 arm_label = "Standard")
# }

# NOT RUN {
add_pr_outcome(ctxml = ctxml,
               name = "correct upload",
               time = "As measured",
               description = "Insert description about the measure.")
# }

# NOT RUN {
add_sec_outcome(ctxml = ctxml,
                name = "time to upload",
                time = "As measured",
                description = "Insert description about the measure.")
# }

# NOT RUN {
add_condition(ctxml = ctxml,
              condition = "Cardiac")
# }

# NOT RUN {
add_keyword(ctxml = ctxml,
            keyword = "sedation")
# }