Print xml document created using spiritR

This function allows you to easily view the structure of the xml document generated using the create_ctxml() function

print_ctxml(ctxml)

Arguments

ctxml
The xml document generated by a call to create_ctxml()

Examples

args_ctxml <- list(
org_name = "UHNToronto",
org_study_id = "Foo trial 20190806",
brief_title = "Foo trial to test auto upload 20190806",
study_acronym = "N/A",
official_title = "Foo trial to test auto upload: A randomized trial new 20190806",
agency  = "Aaron Conway",
resp_party_type = "Sponsor-Investigator",
investigator_username = "aconway",
investigator_title ="Assistant Professor",
brief_summary = "Lay summary here",
start_date = "2019-10",
primary_compl =  "2020-12",
study_compl = "2020-12",
int_subtype = "Health Services Research",
phase = "N/A",
assignment = "Parallel",
allocation = "Randomized",
no_masking =  "False",
masked_subject = "True",
masked_caregiver = "True",
masked_investigator = "True",
masked_assessor = "True",
number_arms = 2,
sample_size = "40",
eligibility_criteria = "Inclusion Criteria
- Adults
Exclusion Criteria
- Children",
healthy_volunteers = "No",
genders_included = "Both",
gender_based = "No",
min_age = "1 years",
max_age = "N/A",
#Central contact
contact_first_name = "Aaron",
contact_last_name = "Conway",
contact_degrees = "PhD",
contact_phone = "649-728-8499",
contact_email = "aaron.conway@utoronto.ca",
#Overall official
official_first_name = "Aaron",
official_last_name ="Conway",
official_degrees = "PhD",
official_affiliation = "UHN",
official_role = "Study Principal Investigator",
#Sharing statements
ipd_sharing = "Yes",
ipd_description = "details",
ipd_protocol =  "True",
ipd_sap = "True",
ipd_icf = "True",
ipd_csr = "True",
ipd_code = "True",
ipd_time = "details",
ipd_criteria = "details",
ipd_url = "http://www.aaronconway.info"
)

ctxml <- do.call(create_ctxml, args_ctxml)

print_ctxml(ctxml)

#> <?xml version="1.0" encoding="UTF-8"?>
#> <study_collection xmlns:prs="http://clinicaltrials.gov/prs">
#>   <clinical_study>
#>     <id_info>
#>       <org_name>UHNToronto</org_name>
#>       <org_study_id>Foo trial 20190806</org_study_id>
#>     </id_info>
#>     <is_fda_regulated/>
#>     <is_section_801/>
#>     <delayed_posting/>
#>     <is_ind_study/>
#>     <brief_title>Foo trial to test auto upload 20190806</brief_title>
#>     <acronym>N/A</acronym>
#>     <official_title>Foo trial to test auto upload: A randomized trial new 20190806</official_title>
#>     <sponsors>
#>       <lead_sponsor>
#>         <agency>Aaron Conway</agency>
#>       </lead_sponsor>
#>       <resp_party>
#>         <resp_party_type>Sponsor-Investigator</resp_party_type>
#>         <investigator_username>aconway</investigator_username>
#>         <investigator_title>Assistant Professor</investigator_title>
#>         <investigator_affiliation/>
#>         <name_title/>
#>         <organization/>
#>         <phone/>
#>         <phone_ext/>
#>         <email/>
#>       </resp_party>
#>     </sponsors>
#>     <oversight_info>
#>       <irb_info>
#>         <approval_status/>
#>       </irb_info>
#>       <has_dmc/>
#>       <fda_regulated_drug/>
#>       <fda_regulated_device/>
#>       <post_prior_to_approval/>
#>     </oversight_info>
#>     <brief_summary>
#>       <textblock>Lay summary here</textblock>
#>     </brief_summary>
#>     <detailed_description>
#>       <textblock/>
#>     </detailed_description>
#>     <overall_status>Not yet recruiting</overall_status>
#>     <why_stopped/>
#>     <verification_date/>
#>     <start_date>2019-10</start_date>
#>     <start_date_type>Anticipated</start_date_type>
#>     <end_date/>
#>     <last_follow_up_date>2020-12</last_follow_up_date>
#>     <last_follow_up_date_type>Anticipated</last_follow_up_date_type>
#>     <prim_compl_date>2020-12</prim_compl_date>
#>     <primary_compl_date_type>Anticipated</primary_compl_date_type>
#>     <study_design>
#>       <study_type>Interventional</study_type>
#>       <interventional_design>
#>         <interventional_subtype>Health Services Research</interventional_subtype>
#>         <phase>N/A</phase>
#>         <assignment>Parallel</assignment>
#>         <model_description>
#>           <textblock/>
#>         </model_description>
#>         <allocation>Randomized</allocation>
#>         <masking/>
#>         <no_masking>False</no_masking>
#>         <masked_subject>True</masked_subject>
#>         <masked_caregiver>True</masked_caregiver>
#>         <masked_investigator>True</masked_investigator>
#>         <masked_assesor>True</masked_assesor>
#>         <masking_description>
#>           <textblock/>
#>         </masking_description>
#>         <control/>
#>         <number_of_arms>2</number_of_arms>
#>       </interventional_design>
#>     </study_design>
#>     <primary_outcome/>
#>     <secondary_outcome/>
#>     <enrollment>40</enrollment>
#>     <enrollment_type>Anticipated</enrollment_type>
#>     <condition/>
#>     <keyword/>
#>     <arm_group/>
#>     <intervention/>
#>     <eligibility>
#>       <study_population>
#>         <textblock/>
#>       </study_population>
#>       <sampling_method/>
#>       <criteria>
#>         <textblock>Inclusion Criteria
#> - Adults
#> Exclusion Criteria
#> - Children</textblock>
#>       </criteria>
#>       <healthy_volunteers>No</healthy_volunteers>
#>       <gender>Both</gender>
#>       <gender_based>No</gender_based>
#>       <gender_description>
#>         <textblock/>
#>       </gender_description>
#>       <minimum_age>1 years</minimum_age>
#>       <maximum_age>N/A</maximum_age>
#>     </eligibility>
#>     <overall_official>
#>       <first_name>Aaron</first_name>
#>       <middle_name/>
#>       <last_name>Conway</last_name>
#>       <degrees>PhD</degrees>
#>       <role>Study Principal Investigator</role>
#>       <affiliation>UHN</affiliation>
#>     </overall_official>
#>     <overall_contact>
#>       <first_name>Aaron</first_name>
#>       <middle_name/>
#>       <last_name>Conway</last_name>
#>       <degrees>PhD</degrees>
#>       <phone>649-728-8499</phone>
#>       <phone_ext/>
#>       <email>aaron.conway@utoronto.ca</email>
#>     </overall_contact>
#>     <overall_contact_backup>
#>       <first_name/>
#>       <middle_name/>
#>       <last_name/>
#>       <degrees/>
#>       <phone/>
#>       <phone_ext/>
#>       <email/>
#>     </overall_contact_backup>
#>     <ipd_sharing_statement>
#>       <sharing_ipd>Yes</sharing_ipd>
#>       <ipd_description>
#>         <textblock>details</textblock>
#>       </ipd_description>
#>       <ipd_info_type_protocol>True</ipd_info_type_protocol>
#>       <ipd_info_type_sap>True</ipd_info_type_sap>
#>       <ipd_info_type_icf>True</ipd_info_type_icf>
#>       <ipd_info_type_csr>True</ipd_info_type_csr>
#>       <ipd_info_type_analytic_code>True</ipd_info_type_analytic_code>
#>       <ipd_time_frame>
#>         <textblock>details</textblock>
#>       </ipd_time_frame>
#>       <ipd_access_criteria>
#>         <textblock>details</textblock>
#>       </ipd_access_criteria>
#>       <ipd_url>http://www.aaronconway.info</ipd_url>
#>     </ipd_sharing_statement>
#>   </clinical_study>
#> </study_collection>